# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daybook digital planner platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inserts:
    get:
      operationId: listInserts
      tags: [inserts]
      summary: List inserts
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, live, all]
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: Inserts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insert"
    post:
      operationId: createInsert
      tags: [inserts]
      summary: Create insert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insert"
  /inserts/{id}:
    get:
      operationId: getInsert
      tags: [inserts]
      summary: Get insert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Insert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insert"
    patch:
      operationId: updateInsert
      tags: [inserts]
      summary: Update insert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insert"
    delete:
      operationId: deleteInsert
      tags: [inserts]
      summary: Delete insert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /inserts/{id}/publish:
    post:
      operationId: publishInsert
      tags: [inserts]
      summary: Publish insert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insert"
  /inserts/{id}/unpublish:
    post:
      operationId: unpublishInsert
      tags: [inserts]
      summary: Unpublish insert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insert"

  # ── RELATED PRODUCTS ─────────────────────────────────────────────────────────
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List subscription plans
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── USERS ────────────────────────────────────────────────────────────────────
...
  /planner-configs:
    get:
      operationId: listPlannerConfigs
      tags: [plannerConfig]
      summary: List planner configs for current user
      responses:
        "200":
          description: Configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannerConfig"
    post:
      operationId: createPlannerConfig
      tags: [plannerConfig]
      summary: Create planner config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerConfigInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerConfig"
  /planner-configs/{id}:
    get:
      operationId: getPlannerConfig
      tags: [plannerConfig]
      summary: Get planner config
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerConfig"
    patch:
      operationId: updatePlannerConfig
      tags: [plannerConfig]
      summary: Update style / output fields (setup fields lock after generation)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerConfig"
    delete:
      operationId: deletePlannerConfig
      tags: [plannerConfig]
      summary: Delete planner config
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── PDF GENERATION ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CatalogStatus:
      type: string
      enum: [draft, live]

    # ── USER ──────────────────────────────────────────────────────────────────
...
    Insert:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        slug:           { type: string }
        description:    { type: ["string", "null"] }
        status:         { $ref: "#/components/schemas/CatalogStatus" }
        category:       { type: ["string", "null"] }
        imageUrl:       { type: ["string", "null"] }
        isTransparent:  { type: boolean }
        driveFileId:    { type: ["string", "null"] }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    InsertInput:
      type: object
      required: [name]
      properties:
        name:           { type: string, minLength: 1 }
        slug:           { type: string }
        description:    { type: string }
        category:       { type: string }
        imageUrl:       { type: string }
        isTransparent:  { type: boolean }
    InsertUpdate:
      type: object
      properties:
        name:           { type: string, minLength: 1 }
        description:    { type: string }
        category:       { type: string }
        imageUrl:       { type: string }
        isTransparent:  { type: boolean }
        driveFileId:    { type: string }

    # ── RELATED PRODUCT ──────────────────────────────────────────────────────
...
    Plan:
      type: object
      required: [id, name, slug, tier, status, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        description:  { type: ["string", "null"] }
        status:       { $ref: "#/components/schemas/CatalogStatus" }
        tier:
          type: string
          enum: [basic, advanced]
        oneTimePrice: { type: ["number", "null"] }
        yearlyPrice:  { type: ["number", "null"] }
        lifetimePrice: { type: ["number", "null"] }
        features:     { type: object, nullable: true, additionalProperties: true }
        stripeProductId: { type: ["string", "null"] }
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
    PlanInput:
      type: object
      required: [name, tier]
      properties:
        name:         { type: string, minLength: 1 }
        slug:         { type: string }
        description:  { type: string }
        tier:
          type: string
          enum: [basic, advanced]
        oneTimePrice: { type: number }
        yearlyPrice:  { type: number }
        lifetimePrice: { type: number }
        features:     { type: object, additionalProperties: true }
    PlanUpdate:
      type: object
      properties:
        name:         { type: string, minLength: 1 }
        description:  { type: string }
        tier:
          type: string
          enum: [basic, advanced]
        oneTimePrice: { type: ["number", "null"] }
        yearlyPrice:  { type: ["number", "null"] }
        lifetimePrice: { type: ["number", "null"] }
        features:     { type: object, nullable: true, additionalProperties: true }
        status:       { $ref: "#/components/schemas/CatalogStatus" }

    # ── PLANNER CONFIG ────────────────────────────────────────────────────────
    PlannerConfig:
      type: object
      required: [id, userId, weekStart, layout, rangeType, createdAt]
      properties:
        id:             { type: integer }
        userId:         { type: integer }
        editionId:      { type: ["integer", "null"] }
        name:           { type: ["string", "null"] }
        # Setup fields (locked after generation)
        weekStart:
          type: string
          enum: [sunday, monday]
        layout:
          type: string
          enum: [2page-landscape, vertical]
        rangeType:
          type: string
          enum: [90day, 12month, monthly, custom]
        startMonth:     { type: ["integer", "null"] }
        startYear:      { type: ["integer", "null"] }
        monthCount:     { type: ["integer", "null"] }
        # Style fields (re-exportable)
        coverColor:     { type: ["string", "null"] }
        coverTexture:   { type: ["string", "null"] }
        accentColor:    { type: ["string", "null"] }
        headingFont:    { type: ["string", "null"] }
        subheadingFont: { type: ["string", "null"] }
        scriptFont:     { type: ["string", "null"] }
        notesSections:
          type: array
          items: { type: string }
        notePaperStyle: { type: ["string", "null"] }
        tabPosition:
          type: ["string", "null"]
          enum: [right, top, bottom, home-only, null]
        calendarLinkType:
          type: ["string", "null"]
          enum: [google, apple-ics, none, null]
        aiPromptBlocks: { type: boolean }
        generatedAt:    { type: ["string", "null"], format: date-time }
        driveUrl:       { type: ["string", "null"] }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    PlannerConfigInput:
      type: object
      required: [weekStart, layout, rangeType]
      properties:
        editionId:      { type: integer }
        name:           { type: string }
        weekStart:
          type: string
          enum: [sunday, monday]
        layout:
          type: string
          enum: [2page-landscape, vertical]
        rangeType:
          type: string
          enum: [90day, 12month, monthly, custom]
        startMonth:     { type: integer, minimum: 1, maximum: 12 }
        startYear:      { type: integer }
        monthCount:     { type: integer, minimum: 1 }
        coverColor:     { type: string }
        coverTexture:   { type: string }
        accentColor:    { type: string }
        headingFont:    { type: string }
        subheadingFont: { type: string }
        scriptFont:     { type: string }
        notesSections:
          type: array
          items: { type: string }
          maxItems: 10
        notePaperStyle: { type: string }
        tabPosition:
          type: string
          enum: [right, top, bottom, home-only]
        calendarLinkType:
          type: string
          enum: [google, apple-ics, none]
        aiPromptBlocks: { type: boolean }
    PlannerConfigUpdate:
      type: object
      properties:
        name:           { type: string }
        coverColor:     { type: string }
        coverTexture:   { type: string }
        accentColor:    { type: string }
        headingFont:    { type: string }
        subheadingFont: { type: string }
        scriptFont:     { type: string }
        notesSections:
          type: array
          items: { type: string }
          maxItems: 10
        notePaperStyle: { type: string }
        tabPosition:
          type: string
          enum: [right, top, bottom, home-only]
        calendarLinkType:
          type: string
          enum: [google, apple-ics, none]
        aiPromptBlocks: { type: boolean }

    # ── GENERATION ───────────────────────────────────────────────────────────
...